requirements.txt
fastapi>=0.112.0
uvicorn>=0.30.0
google-genai>=0.3.0
notion-client>=2.2.1
python-dotenv>=1.0.1
pydantic>=2.8.0